% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{check_package}
\alias{check_package}
\title{Return error message and abort if package not found}
\usage{
check_package(package)
}
\arguments{
\item{package}{The name of the package that to be installed}
}
\description{
Return error message and abort if package not found
}
