% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{col_search}
\alias{col_search}
\title{Search for column names that match given string}
\usage{
col_search(df, search_string, ignore_case = TRUE)
}
\arguments{
\item{df}{Dataframe to search in}

\item{search_string}{String to search column names for}

\item{ignore_case}{Should case be ignored in column search. Defaults to TRUE}
}
\value{
A character vector of column names that match search
}
\description{
Search for column names that match given string
}
