% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.r
\name{colMaxs}
\alias{colMaxs}
\title{colMaxs}
\usage{
colMaxs(mat, ...)
}
\arguments{
\item{mat}{Matrix containing numeric values}

\item{...}{Additional parameters to be passed to max}
}
\description{
Computes the maximum value in each column for a given matrix.
}
