% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.r
\name{concat}
\alias{concat}
\alias{\%cat\%}
\title{quick concat}
\usage{
x \%cat\% y
}
\arguments{
\item{x}{Variable to concatenate first}

\item{y}{Variable to concatenate second}
}
\description{
quickly concatenate two variables
}
\examples{
1 \%cat\% 2
}
