% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.r
\name{count_lines}
\alias{count_lines}
\title{Count Lines}
\usage{
count_lines(filepath, chunk = 2^20)
}
\arguments{
\item{filepath}{A path to the text file to read}

\item{chunk}{The number of bytes to read in at one time}
}
\value{
The number of lines
}
\description{
Count the number of lines in a given text file
}
\details{
This function counts newline bytes in chunks to avoid reading the entire file into memory at once. A final line
without a trailing newline is counted. Gzip, bzip2 and xz compressed files are read transparently.
}
