% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{create_dir}
\alias{create_dir}
\title{Create dir if it doesn't exist}
\usage{
create_dir(name)
}
\arguments{
\item{name}{Name of directory}
}
\description{
If supplied a filepath with multiple directories will create all parent directories
}
