% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_setup.r
\name{create_lintr}
\alias{create_lintr}
\title{Make an .lintr file in specified directory}
\usage{
create_lintr(path = ".")
}
\arguments{
\item{path}{The path to place the .lintr file}
}
\value{
TRUE if file was successfully created
}
\description{
Make an .lintr file in specified directory
}
