% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{create_list}
\alias{create_list}
\title{create_list}
\usage{
create_list(names)
}
\arguments{
\item{names}{A character or integer vector of names to be assigned to list.}
}
\value{
An empty named list of length \code{names}.
}
\description{
Create an empty list that is length \code{names} and with each element named by the vector of names provided. This is
shorthand for creating an empty list and then assigning it names.
}
