% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.r
\name{diversity}
\alias{diversity}
\title{Diversity}
\usage{
diversity(matrix, index = "shannon")
}
\arguments{
\item{matrix}{The matrix containing counts}

\item{index}{Which diversity index to calculat}
}
\value{
A vector of indices for each row of the matrix provided
}
\description{
Calculate a number of diversity indices
}
\details{
This function assumes that rows are independent samples and columns are counts of differing species. For some
diversity measures the input matrix must be positive and finite, if the output is NaN this is likely the issue.
}
