% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{estimate_plot}
\alias{estimate_plot}
\title{Return a plot of a list of models estimates and their intervals}
\usage{
estimate_plot(model, model_names = names(model), coefficient = NULL, ...)
}
\arguments{
\item{model}{A list of statistical model objects, or a single model.}

\item{model_names}{A character vector of model names. Defaults to the names of \code{model}.}

\item{coefficient}{The model term to be compared across all models.}

\item{...}{Additional arguments passed to broom::tidy.}
}
\value{
A ggplot of all model estimates
}
\description{
Creates a scatter plot with one point for each estimate and lineranges for their confidence intervals.
}
