% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_colref}
\alias{get_colref}
\title{get_colref}
\usage{
get_colref(df, x)
}
\arguments{
\item{df}{The dataframe to evaluate}

\item{x}{The column name (in string format) or column number (integer format) that you wish to evaluate}
}
\description{
Returns the column number that corresponds to a given column name or vice-versa if a number is supplied
}
