% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Return a grid of plots that share a legend}
\usage{
grid_arrange_shared_legend(
  ...,
  ncol = length(list(...)),
  nrow = 1,
  position = c("bottom", "right"),
  plot_legend = 1
)
}
\arguments{
\item{...}{The ggplot2 objects to plot}

\item{ncol}{The number of columns to arrange the plots into.}

\item{nrow}{The number of rows to arrange the plots into.}

\item{position}{Where to place the shared legend in the grid of plots, either at the bottom or right.}

\item{plot_legend}{Which plot to pull the legend element from, defaults to the first provided plot.}
}
\value{
An invisible plot containing all supplied ggplots, as well as a shared legend located at \code{position}.
}
\description{
Creates a n x n grid of plots that all share the same legend using grid.arrange.
}
