% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{is_date}
\alias{is_date}
\title{is_date}
\usage{
is_date(x)
}
\arguments{
\item{x}{Vector you wish to check is a date.}
}
\description{
Checks to see if the given vector is one of the two base R date objects (Date, POSIXt)
}
