% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.r
\name{is_orthogonal}
\alias{is_orthogonal}
\title{is_orthogonal}
\usage{
is_orthogonal(x, coerce = FALSE)
}
\arguments{
\item{x}{Square/rectangular matrix you wish to evaluate.}

\item{coerce}{Boolean indicating whether to attempt to coerce the input into a square matrix}
}
\description{
Returns TRUE/FALSE is a matrix is orthogonal.
}
