% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{lumiscversion}
\alias{lumiscversion}
\title{Installed version of R/lumisc}
\usage{
lumiscversion()
}
\value{
A character string with the version number of the currently installed
version of R/lumisc.
}
\description{
Print the version number of the currently installed version of R/lumisc.
}
\examples{
lumiscversion()
}
\keyword{print}
