% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{mode}
\alias{mode}
\title{Mode}
\usage{
mode(x)
}
\arguments{
\item{x}{Vector you wish to evaluate.}
}
\description{
Returns the mode (number that occurs the most) of a given vector. All tied values are returned, and \code{NA} counts as
a value.
}
