% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.r
\name{nil}
\alias{nil}
\alias{\%||\%}
\title{Default value for NULL}
\usage{
lhs %||% rhs
}
\description{
Return \code{lhs} if \code{lhs} is not \code{NULL} otherwise returns
code{rhs}
}
\examples{
1 \%||\% 2 # returns 1
NULL \%||\% 2 # returns 2
}
