% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nin}
\alias{nin}
\alias{\%!in\%}
\title{Inverse Value Matching}
\usage{
x %in% c(x, y, z)
}
\description{
Complement of \code{\%in\%}. Returns the elements of \code{x} that are
not in \code{y}.
}
\examples{
1 \%!in\% 1:10
1 \%!in\% 2:10
11 \%!in\% 1:10
}
