% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{q}
\alias{q}
\alias{quit}
\title{quit}
\usage{
q(save = "no", ...)
}
\arguments{
\item{save}{Default \code{no}. Choose whether or not to save the workspace.}

\item{...}{Additional arguments passed to the \code{quit()} function.}
}
\description{
Quits R without saving the workspace to an image
}
