% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{rgb2hex}
\alias{rgb2hex}
\title{Return hex value of specified rgb colour}
\usage{
rgb2hex(r, g, b)
}
\arguments{
\item{r}{R channel}

\item{g}{G channel}

\item{b}{B channel}
}
\value{
A hex colour code
}
\description{
Take an RGB colour value and return a hex colour code
}
