% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_options.r
\name{set_startup_options}
\alias{set_startup_options}
\title{Set Nice Startup Options}
\usage{
set_startup_options(
  digits = 4L,
  bitmapType = "cairo",
  show.signif.stars = FALSE,
  useFancyQuotes = FALSE,
  width = 88L,
  Ncpus = max(1L, parallel::detectCores() - 1L, na.rm = TRUE),
  max.print = 100L,
  servr.daemon = TRUE,
  mc.cores = max(1L, parallel::detectCores() - 1L, na.rm = TRUE),
  error = "rlang",
  menu.graphics = FALSE,
  continue = "-_- ",
  warnPartialMatchArgs = TRUE,
  warnPartialMatchDollar = TRUE,
  warnPartialMatchAttr = TRUE,
  nwarnings = 1e+06,
  scipen = 999L,
  datatable.print.class = TRUE,
  browser = "xdg-open",
  HTTPUserAgent = sprintf("R/\%s R (\%s)", getRversion(), paste(getRversion(),
    R.version$platform, R.version$arch, R.version$os)),
  download.file.extra = sprintf("--header \\"User-Agent: R (\%s)\\"",
    paste(getRversion(), R.version$platform, R.version$arch, R.version$os)),
  ...
)
}
\arguments{
\item{digits}{Default \code{4}}

\item{bitmapType}{Default "cairo".}

\item{show.signif.stars}{Default \code{FALSE}}

\item{useFancyQuotes}{Default \code{FALSE}}

\item{width}{Default \code{88}}

\item{Ncpus}{Default number of CPUs - 1. Used for parallel pkg installs.}

\item{max.print}{Default 100 to avoid blow up}

\item{servr.daemon}{Default \code{TRUE}. For xaringan presentations}

\item{mc.cores}{Default number of CPUs - 1. Used for parallel computing}

\item{error}{Default \code{"rlang"}, which sets \code{rlang::entrace}. Otherwise passed to \code{options} as is.}

\item{menu.graphics}{Default \code{FALSE}. Logical: should graphical menus be used if available?}

\item{continue}{Default \code{"-_- "}. Set the prompt used for lines which continue over one line.}

\item{warnPartialMatchArgs}{Default \code{TRUE}. Warn if using partial arguments.}

\item{warnPartialMatchDollar}{Default \code{TRUE}. Warns if partial matching is used for extraction by $.}

\item{warnPartialMatchAttr}{Default \code{TRUE}. Warns if partial matching is used to extract attributes via attr.}

\item{nwarnings}{Default \code{1e6}.}

\item{scipen}{Default \code{999}. Always print out full numbers, i.e. not 1e2}

\item{datatable.print.class}{Default \code{TRUE}. Always print data.table column class.}

\item{browser}{Default \code{xdg-open}. Browser to open http help documents.}

\item{HTTPUserAgent}{Used by RStudio Package Manager (RSPM).}

\item{download.file.extra}{Used by RSPM for curl/wget installs, e.g. Rscript.}

\item{...}{Other arguments passed to \code{options}.}
}
\description{
Sets nicer options. All arguments are passed to the \code{options} function.
This function also sets \code{ipck = TRUE} in \code{rc.settings}.
}
