% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{view}
\alias{view}
\title{View}
\usage{
view(x, title)
}
\arguments{
\item{x}{An R object which can be coerced to a data frame with non-zero numbers of rows and columns.}

\item{title}{Title for viewer window.  Defaults to name of 'x' prefixed by 'Data:'.}
}
\description{
Invoke a spreadsheet-style data viewer on a matrix-like R object.
}
