% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{view_last}
\alias{view_last}
\title{view_last}
\usage{
view_last(title)
}
\arguments{
\item{title}{Title for viewer window. Defaults to name of 'x' prefixed by 'Data:'.}
}
\description{
Invoke a spreadsheet-style data viewer on the last R object interacted with.
}
