% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.r
\name{write_df}
\alias{write_df}
\title{Write Dataframes}
\usage{
write_df(list, file, overwrite = TRUE, ...)
}
\arguments{
\item{list}{The list of dataframes to append to a single csv file.}

\item{file}{File name of the final csv.}

\item{overwrite}{Boolean value, if true will always overwrite file even if it exists, otherwise appends to it.}

\item{...}{Additional options passed to \code{\link[data.table:fwrite]{data.table::fwrite()}}.}
}
\description{
Write a list of rectangular dataframes to a single csv keeping column names, separating dataframes with a blank line.
}
\details{
This function never writes row names to the csv file.
}
