% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cache_file}
\alias{cache_file}
\title{cache_file}
\usage{
cache_file(file, expr, force = FALSE, ...)
}
\arguments{
\item{file}{Path to the cache file.}

\item{expr}{The expression to evaluate if the cache doesn't exist.}

\item{force}{Logical; if TRUE, re-run the expression even if the file exists.}

\item{...}{Additional arguments passed to saveRDS.}
}
\description{
cache_file
}
