% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.r
\name{colVars}
\alias{colVars}
\title{colVars}
\usage{
colVars(mat, coerce = FALSE)
}
\arguments{
\item{mat}{Matrix containing numeric values}

\item{coerce}{Boolean indicating whether to attempt to coerce the input into a matrix}
}
\description{
Computes the column wise variances on a given matrix.
}
