% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.r
\name{colVars}
\alias{colVars}
\title{colVars}
\usage{
colVars(mat, ...)
}
\arguments{
\item{mat}{Matrix containing numeric values}

\item{...}{Additional parameters to be passed to variance function}
}
\description{
Computes the column wise variances on a given matrix.
}
