% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.r
\name{count_lines}
\alias{count_lines}
\title{Count Lines}
\usage{
count_lines(filepath, batch)
}
\arguments{
\item{filepath}{A path to the text file to read}

\item{batch}{The number of lines to read in at one time}
}
\value{
An integer indicating number of lines
}
\description{
Count the number of lines in a given text file
}
\details{
This function counts the lines of a given text file line by line to prevent reading the entire file into memory at
once. Higher batch numbers generally equal quicker read times.
}
