% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_setup.r
\name{create_air}
\alias{create_air}
\title{Make an air.toml file in specified directory}
\usage{
create_air(path)
}
\arguments{
\item{path}{The path to place the air.toml file}
}
\value{
TRUE if file was successfully created
}
\description{
Make an air.toml file in specified directory
}
