% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_setup.r
\name{create_rproj}
\alias{create_rproj}
\title{Make an .rproj file in specified directory}
\usage{
create_rproj(path)
}
\arguments{
\item{path}{The name of the .Rproj file to create}
}
\value{
TRUE if file was successfully created
}
\description{
Make an .rproj file in specified directory
}
