% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.r
\name{donut_geomask}
\alias{donut_geomask}
\title{Donut Geomasking}
\usage{
donut_geomask(xcoords, ycoords, min_dist = 5, max_dist = 10, units = "m")
}
\arguments{
\item{xcoords}{X coordinates to transform}

\item{ycoords}{Y coordinates to transform}

\item{min_dist}{The minimum distance the new transformed points must be from the original point}

\item{max_dist}{The maximum distance the new transformed points must be from the original point}

\item{units}{The units to do the transformation in by default meters, else will perform transformation on coordinate
data}
}
\value{
A list containing new x and y coordinate vectors
}
\description{
Transform x:y coordinates to new random x:y coordinates within a defined donut to mask original location.
This function works with GPS coordinates as well. Work is being done on supporting lon:lat coordinates.
}
\author{
Maia Lesosky

Luke Hannan
}
