% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{estimate_plot}
\alias{estimate_plot}
\title{Return a plot of a list of models estimates and their intervals}
\usage{
estimate_plot(model_list, model_names, coefficient = NULL)
}
\arguments{
\item{model_list}{A list of statistical model objects.}

\item{model_names}{A character vector containing the list of model names.}

\item{coefficient}{The model term to be compared across all models.}
}
\value{
An invisible plot containing all model estimates as a ggplot
}
\description{
Creates a scatter plot with one point for each estimate and lineranges for their confidence intervals.
}
