% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumisc.r
\docType{package}
\name{lumisc}
\alias{lumisc}
\alias{_PACKAGE}
\alias{lumisc-package}
\title{Personal Collection of Miscellaneous Functions}
\description{
A package of collected miscellaneous functions I use regularly, and
may prove useful.
}
\details{
A package of collected miscellaneous functions I use regularly, and
may prove useful. Otherwise refer to the vignettes to see how to format the
documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/luhann/lumisc}
}

}
\author{
\strong{Maintainer}: Luke Hannan \email{luke@lukehannan.com}

}
\keyword{internal}
