% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nil}
\alias{nil}
\alias{\%||\%}
\title{Default value for NULL}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{An object to set if it's null}

\item{rhs}{The value to provide if x is null}
}
\description{
Return \code{lhs} if \code{lhs} is not \code{NULL} otherwise returns
code{rhs}
}
\examples{
1 \%||\% 2
NULL \%||\% 2
}
