% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.r
\name{\%nin\%}
\alias{\%nin\%}
\title{Inverse Value Matching}
\usage{
x %in% c(x, y, z)
}
\description{
Complement of \code{\%in\%}. Returns the elements of \code{x} that are
not in \code{y}.
}
\examples{
 1 \%nin\% 1:10
 1 \%nin\% 2:10
}
