% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.r
\name{retrieve_credentials}
\alias{retrieve_credentials}
\title{Read a token and other credentials from a file}
\usage{
retrieve_credentials(
  path_credential,
  project_id = NA_character_,
  check_url = FALSE,
  username = NA_character_,
  uri_name = NA_character_
)
}
\arguments{
\item{path_credential}{The path to the credentials file.}

\item{project_id}{Allows project_id to be specified, particularly useful when interacting with the redcap api.}

\item{check_url}{Should the urls in the credentials file be validated. Defaults to FALSE.}

\item{username}{Username to search for in credentials file.}

\item{uri_name}{uri of project to search for in credentials file.}
}
\description{
This function helps manage tokens securely for connecting to various online data sources.
}
