% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.r
\name{splitn}
\alias{splitn}
\title{splitn}
\usage{
splitn(mat, r = 1, c = ncol(mat))
}
\arguments{
\item{mat}{Matrix containing numeric values}

\item{r}{Rows in each submatrix}

\item{c}{Cols in each submatrix}
}
\description{
Splits a given matrix into a list of matrices (byrow) into a list of submatrices
}
