% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{switchv}
\alias{switchv}
\title{switchv}
\usage{
switchv(EXPR, ...)
}
\arguments{
\item{EXPR}{An expression evaluating to a vector of numbers of strings}

\item{...}{List of alternatives}
}
\value{
Vector of returned values, simplified to their common type. Elements with no matching alternative are \code{NA}.
}
\description{
Vectorized version of switch
}
\details{
Vectorized version of \code{\link[base:switch]{base::switch()}}: calls \code{\link[base:switch]{base::switch()}} once per unique
value of the input and maps the results back.
}
\examples{
switchv(c("horse", "fish", "cat", "bug"),
  horse = "fast",
  cat = "cute",
  "what?"
)
}
\author{
Karl Broman
}
